requirements.txt
Flask==3.1.0
gunicorn==23.0.0
gspread==6.1.2
google-auth==2.35.0
python-dotenv==1.0.1